Consolidate common argument validation code for array segments
